Final variant

+++ src/put_di.c
@@ -6,7 +6,7 @@
 /*   By: vbrazas <vbrazas@student.unit.ua>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/03 12:38:02 by vbrazas           #+#    #+#             */
-/*   Updated: 2018/01/13 19:03:10 by vbrazas          ###   ########.fr       */
+/*   Updated: 2018/01/16 14:53:07 by vbrazas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,13 +76,13 @@ static	int		handle_minln(char *s, intmax_t n, unsigned int i, t_flags *fl)
 		while (i > ((n < 0 || fl->plus || fl->space)\
 		&& fl->zero && !fl->precs_spec && !fl->minus) ? 1 : 0)
 			ml[--j] = s[--i];
-		j = fl->minus ? ft_strlen(s) : 0;
+		j = fl->minus ? (unsigned int)ft_strlen(s) : 0;
 		while (j < fl->min_lenth && fl->minus)
 			ml[j++] = ' ';
 		free(s);
 		s = ml;
 	}
-	j = ft_putstr(s);
+	j = (unsigned int)ft_putstr(s);
 	free(s);
 	free(fl);
 	return (j);
@@ -110,7 +110,7 @@ static	int		help(char *s, intmax_t n, unsigned int i, t_flags *fl)
 		free(s);
 		s = precision;
 	}
-	i = ft_strlen(s);
+	i = (unsigned int)ft_strlen(s);
 	return (handle_minln(s, n, i, fl));
 }
 
@@ -122,6 +122,6 @@ int				put_di(intmax_t n, t_flags *fl)
 	s = estoa_base(n, fl->base, fl);
 	if (fl->precs_spec && !fl->precision && !n)
 		s[0] = 0;
-	i = ft_strlen(s);
+	i = (unsigned int)ft_strlen(s);
 	return (help(s, n, i, fl));
 }
